pyproject.toml
[project]
name = "varlabzbot"
version = "0.1.0"
description = "varlabz telegram bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ask",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
]

[tool.uv.sources]
ask = { git = "ssh://git@github.com/varlabz/ask.git" }
